refactor: 전체 상수 적용
