+++ src/main/java/blackjack/model/BlackjackConfig.java
@@ -1,0 +1,15 @@
+package blackjack.model;
+
+public class BlackjackConfig {
+
+    public static final int DEALER_STAND_SCORE = 17;
+    public static final int BLACKJACK_SCORE = 21;
+
+    public static final int INIT_CARDS_START = 0;
+    public static final int INITIAL_CARD_COUNT = 2;
+
+    public static final int ACE_LOW_SCORE = 1;
+    public static final int ACE_HIGH_SCORE = 11;
+
+}
+
