--- src/main/java/blackjack/model/Constant.java
@@ -1,15 +1,0 @@
-package blackjack.model;
-
-public class Constant {
-
-    public static final int DEALER_ADD_CARD_STAND = 17;
-    public static final int BLACKJACK_SCORE = 21;
-
-    public static final int INIT_CARDS_START_IDX = 0;
-    public static final int INITIAL_CARD_COUNT = 2;
-
-    public static final int ACE_SCORE_ONE = 1;
-    public static final int ACE_SCORE_ELEVEN = 11;
-
-}
-
